Migrate list-item to TypeScript
