+++ src/window_main/history.js
@@ -13,7 +13,7 @@ import { makeId } from "../shared/util";
 import Aggregator from "./aggregator";
 import DataScroller from "./data-scroller";
 import FilterPanel from "./FilterPanel";
-import ListItem from "./list-item";
+import ListItem from "./listItem";
 import StatsPanel from "./stats-panel";
 import {
   attachDraftData,
